pr_code_reviewer/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-adk
google-genai
PyGithub
requests
